uploaded graph api class
This is a class I created for internal use.  As we deprecated IMAP company wide, a new solution was required for all of our scripts that used IMAPlib to catch and download attachments from our customer facing email address, digital.process@.  I used the new MS Graph API along with python, and our Outlook environment to create a class with methods that supplanted IMAPlib internally, and enabled continuity.